202507251128/requirements.txt
cryptography==45.0.5
PyMuPDF==1.26.3
keyring==25.6.0
ocrmypdf==16.10.4
pdf2image==1.17.0
Pillow==11.3.0
pyodbc==5.2.0
PyPDF2==3.0.1
pytesseract==0.3.13
pywin32==310
reportlab==4.4.3
requests==2.32.4
watchdog==6.0.0
python-docx==1.2.0
openpyxl==3.1.5
lxml==6.0.0
python-dateutil==2.8.2
